hello_world/requirements.txt
requests
selenium
beautifulsoup4
Pillow
webdriver-manager